+++ R/chTKCat.R
@@ -9,14 +9,20 @@
 #' @param host a character string specifying the host heberging the 
 #' database (default: localhost)
 #' @param port an integer specifying the port on which the 
-#' database is listening (default: 9101)
+#' database is listening (default: 9111)
 #' @param user user name
 #' @param password user password
-#' @param http an integer specifying the HTTP port of the 
-#' ClickHouse database (default: NULL). Used for documentation only.
 #' @param settings list of
 #' [Clickhouse 
 #' settings](https://clickhouse.com/docs/en/operations/settings/settings/)
+#' @param ports a named list of available ports for accessing ClickHouse
+#' (default: NULL; example: `c(Native=9101, HTTP=9111)`)
+#' @param drv a DBI driver for connecting to ClickHouse
+#' (default: [ClickHouseHTTP::ClickHouseHTTP()];
+#' other supported driver: [RClickhouse::clickhouse()])
+#' @param ... additional parameters for connection
+#' (see [ClickHouseHTTP::dbConnect,ClickHouseHTTPDriver-method]
+#' for the default driver)
 #'
 #' @return a chTKCat object
 #'
@@ -26,10 +32,9 @@
 #'
 chTKCat <- function(
    host="localhost",
-   port=9101L,
+   port=9111L,
    user="default",
    password,
-   http=NULL,
    settings=list(
       # Which part of the query can be read into RAM for parsing
       # (the remaining data for INSERT, if any, is read later)
@@ -47,31 +52,49 @@ chTKCat <- function(
       # Force joined subqueries and table functions to have aliases for correct
       # name qualification. Zero means FALSE.
       "joined_subquery_requires_alias"=0
-   )
+   ),
+   ports=NULL,
+   drv=ClickHouseHTTP::ClickHouseHTTP(),
+   ...
 ){
+   
+   if(!class(drv) %in% c("ClickhouseDriver", "ClickHouseHTTPDriver")){
+      stop(
+         "Only ClickhouseDriver from RClickhouse package",
+         " and ClickHouseHTTPDriver from ClickHouseHTTP package",
+         " are supported."
+      )
+   }
+   
    if(missing(password)){
       password <- getPass::getPass(
          sprintf("%s password (press escape or cancel if no password)", user),
          noblank=TRUE
       )
+      if(is.null(password)){
+         password=""
+      }
    }
    
-   chcon <- RClickhouse::dbConnect(
-      drv=RClickhouse::clickhouse(),
+   chcon <- DBI::dbConnect(
+      drv=drv,
       host=host,
       port=port,
-      user=user, password=ifelse(is.na(password), "", password)
+      user=user, password=ifelse(is.na(password), "", password),
+      ...
    )
    
    
    for(s in names(settings)){
-      RClickhouse::dbSendQuery(chcon, sprintf("SET %s='%s'", s, settings[[s]]))
+      DBI::dbSendQuery(chcon, sprintf("SET %s='%s'", s, settings[[s]]))
    }
    
    toRet <- list(
       chcon=chcon,
-      http=http,
-      settings=settings
+      settings=settings,
+      ports=ports,
+      drv=drv,
+      cpar=list(...)
    )
    class(toRet) <- "chTKCat"
    
@@ -170,10 +193,10 @@ check_chTKCat <- function(x, verbose=FALSE){
          stop("Not a chTKCat")
       }
       toRet$init <- TRUE
-      toRet$instance <- dbSys$instance
-      toRet$version <- dbSys$version
-      toRet$contact <- dbSys$contact
-      toRet$path <- dbSys$path
+      toRet$instance <- dbSys[["instance"]]
+      toRet$version <- dbSys[["version"]]
+      toRet$contact <- dbSys[["contact"]]
+      toRet$path <- dbSys[["path"]]
       toRet$admin <- admin
       toRet$provider <- provider
    }
@@ -296,7 +319,7 @@ print.chTKCat <- function(x, ...){
 #' @export
 #'
 db_disconnect.chTKCat <- function(x){
-   RClickhouse::dbDisconnect(x[["chcon"]])
+   DBI::dbDisconnect(x[["chcon"]])
    invisible()
 }
 
@@ -310,12 +333,63 @@ db_disconnect.chTKCat <- function(x){
 db_reconnect.chTKCat <- function(x, user, password, ntries=3, ...){
    xn <- deparse(substitute(x))
    con <- x$chcon
-   db_reconnect(
-      con, user=user, password=password, ntries=ntries,
-      settings=x$settings
-   )
+   
+   if(missing(user)){
+      user <- con@user
+   }
+   suppressWarnings(try(DBI::dbDisconnect(con), silent=TRUE))
+   if(missing(password)){
+      ncon <- try(do.call(DBI::dbConnect, c(
+         list(
+            drv=x$drv,
+            host=con@host,
+            port=con@port,
+            user=user,
+            password=""
+         ),
+         x$cpar
+      )), silent=TRUE)
+      n <- 0
+      while(inherits(ncon, "try-error") & n < ntries){
+         password <- getPass::getPass(
+            msg=paste0(user, " password on ", con@host, ":", con@port)
+         )
+         if(is.null(password)){
+            stop("Canceled by the user")
+         }
+         ncon <- try(do.call(DBI::dbConnect, c(
+            list(
+               drv=x$drv,
+               host=con@host,
+               port=con@port,
+               user=user,
+               password=password
+            ),
+            x$cpar
+         )), silent=TRUE)
+         n <- n+1
+      }
+      if(inherits(ncon, "try-error")){
+         stop(as.character(ncon))
+      }
+   }else{
+      ncon <- try(do.call(DBI::dbConnect, c(
+         list(
+            drv=x$drv,
+            host=con@host,
+            port=con@port,
+            user=user,
+            password=password
+         ),
+         x$cpar
+      )), silent=TRUE)
+   }
+   for(s in names(x$settings)){
+      DBI::dbSendQuery(ncon, sprintf("SET %s='%s'", s, x$settings[[s]]))
+   }
+   
    nv <- x
-   nv$chcon <- con
+   nv$chcon <- ncon
    nv <- check_chTKCat(nv)
    assign(xn, nv, envir=parent.frame(n=1))
    invisible(nv)
@@ -454,13 +528,18 @@ init_chTKCat <- function(
    xml2::xml_replace(users, xml2::read_xml("<users></users>"))
    xml2::write_xml(uf, file=TKCAT_USERS)
    Sys.sleep(3)
-   x <- chTKCat(
-      host=con@host,
-      port=con@port,
-      http=x$http,
-      user=login,
-      password=password
-   )
+   x <- do.call(chTKCat, c(
+      list(
+         host=con@host,
+         port=con@port,
+         user=login,
+         password=password,
+         settings=x$settings,
+         ports=x$ports,
+         drv=x$drv
+      ),
+      x$cpar
+   ))
    con <- x$chcon
    
    ## Create default user ----
@@ -561,7 +640,7 @@ create_chTKCat_user <- function(
    )
    con <- x$chcon
    ## Create the user in ClickHouse ----
-   RClickhouse::dbSendQuery(
+   DBI::dbSendQuery(
       con, 
       sprintf(
          "CREATE USER %s %s",
@@ -589,7 +668,7 @@ create_chTKCat_user <- function(
    
    ## Grant access ----
    if(admin){
-      RClickhouse::dbSendQuery(
+      DBI::dbSendQuery(
          con,
          sprintf(
             "GRANT ALL ON *.* TO %s WITH GRANT OPTION",
@@ -598,7 +677,7 @@ create_chTKCat_user <- function(
       )
    }else{
       
-      RClickhouse::dbSendQuery(
+      DBI::dbSendQuery(
          con,
          sprintf(
             "REVOKE ALL ON *.* FROM %s",
@@ -607,7 +686,7 @@ create_chTKCat_user <- function(
       )
       
       if(provider){
-         RClickhouse::dbSendQuery(
+         DBI::dbSendQuery(
             con, 
             sprintf(
                paste(
@@ -619,14 +698,14 @@ create_chTKCat_user <- function(
             )
          )
          
-         RClickhouse::dbSendQuery(
+         DBI::dbSendQuery(
             con,
             sprintf(
                "REVOKE ALL ON default.* FROM %s",
                login
             )
          )
-         RClickhouse::dbSendQuery(
+         DBI::dbSendQuery(
             con,
             sprintf(
                "REVOKE ALL ON system.* FROM %s",
@@ -636,16 +715,16 @@ create_chTKCat_user <- function(
          
       }
       
-      RClickhouse::dbSendQuery(
+      DBI::dbSendQuery(
          con, sprintf("GRANT SHOW DATABASES ON *.* TO %s", login)
       )
-      RClickhouse::dbSendQuery(
+      DBI::dbSendQuery(
          con, sprintf("GRANT SHOW TABLES ON *.* TO %s", login)
       )
-      RClickhouse::dbSendQuery(
+      DBI::dbSendQuery(
          con, sprintf("GRANT SHOW COLUMNS ON *.* TO %s", login)
       )
-      RClickhouse::dbSendQuery(
+      DBI::dbSendQuery(
          con,
          sprintf(
             paste(
@@ -655,10 +734,10 @@ create_chTKCat_user <- function(
             login
          )
       )
-      RClickhouse::dbSendQuery(
+      DBI::dbSendQuery(
          con, sprintf("GRANT SELECT ON default.Collections TO %s", login)
       )
-      RClickhouse::dbSendQuery(
+      DBI::dbSendQuery(
          con,
          sprintf(
             "GRANT SELECT(login, admin, provider) ON default.Users TO %s",
@@ -707,7 +786,7 @@ change_chTKCat_password <- function(
    )
    con <- x$chcon
    ## Alter the user in ClickHouse ----
-   RClickhouse::dbSendQuery(
+   DBI::dbSendQuery(
       con, 
       sprintf(
          "ALTER USER %s %s",
@@ -789,7 +868,7 @@ update_chTKCat_user <- function(
       admin <- new_val$admin
 
       ## Update the value in 2 steps because of issues with Clickhouse UPDATE 
-      RClickhouse::dbSendQuery(
+      DBI::dbSendQuery(
          con,
          sprintf(
             "ALTER TABLE default.Users DELETE WHERE login='%s'",
@@ -801,7 +880,7 @@ update_chTKCat_user <- function(
       ## Update GRANTs if necessary ----
       if(updateGrants){
          if(admin){
-            RClickhouse::dbSendQuery(
+            DBI::dbSendQuery(
                con,
                sprintf(
                   "GRANT ALL ON *.* TO %s WITH GRANT OPTION",
@@ -810,7 +889,7 @@ update_chTKCat_user <- function(
             )
          }else{
             
-            RClickhouse::dbSendQuery(
+            DBI::dbSendQuery(
                con,
                sprintf(
                   "REVOKE ALL ON *.* FROM %s",
@@ -819,7 +898,7 @@ update_chTKCat_user <- function(
             )
             
             if(provider){
-               RClickhouse::dbSendQuery(
+               DBI::dbSendQuery(
                   con, 
                   sprintf(
                      paste(
@@ -832,14 +911,14 @@ update_chTKCat_user <- function(
                   )
                )
                
-               RClickhouse::dbSendQuery(
+               DBI::dbSendQuery(
                   con,
                   sprintf(
                      "REVOKE ALL ON default.* FROM %s",
                      login
                   )
                )
-               RClickhouse::dbSendQuery(
+               DBI::dbSendQuery(
                   con,
                   sprintf(
                      "REVOKE ALL ON system.* FROM %s",
@@ -849,16 +928,16 @@ update_chTKCat_user <- function(
                
             }
             
-            RClickhouse::dbSendQuery(
+            DBI::dbSendQuery(
                con, sprintf("GRANT SHOW DATABASES ON *.* TO %s", login)
             )
-            RClickhouse::dbSendQuery(
+            DBI::dbSendQuery(
                con, sprintf("GRANT SHOW TABLES ON *.* TO %s", login)
             )
-            RClickhouse::dbSendQuery(
+            DBI::dbSendQuery(
                con, sprintf("GRANT SHOW COLUMNS ON *.* TO %s", login)
             )
-            RClickhouse::dbSendQuery(
+            DBI::dbSendQuery(
                con,
                sprintf(
                   paste(
@@ -868,10 +947,10 @@ update_chTKCat_user <- function(
                   login
                )
             )
-            RClickhouse::dbSendQuery(
+            DBI::dbSendQuery(
                con, sprintf("GRANT SELECT ON default.Collections TO %s", login)
             )
-            RClickhouse::dbSendQuery(
+            DBI::dbSendQuery(
                con,
                sprintf(
                   "GRANT SELECT(login, admin, provider) ON default.Users TO %s",
@@ -919,11 +998,11 @@ drop_chTKCat_user <- function(x, login){
    for(mdb in allDb){
       remove_chMDB_user(x=x, login=login, mdb=mdb)
    }
-   RClickhouse::dbSendQuery(
+   DBI::dbSendQuery(
       con,
       sprintf("ALTER TABLE default.Users DELETE WHERE login='%s'", login)
    )
-   RClickhouse::dbSendQuery(
+   DBI::dbSendQuery(
       con,
       sprintf("DROP USER %s", login)
    )
@@ -1226,7 +1305,7 @@ create_chMDB <- function(x, name, public=FALSE){
       stop("The database exists already")
    }
    con <- x$chcon
-   RClickhouse::dbSendQuery(con, sprintf("CREATE DATABASE `%s`", name))
+   DBI::dbSendQuery(con, sprintf("CREATE DATABASE `%s`", name))
    mergeTrees_from_RelDataModel(
       con, name,
       CHMDB_DATA_MODEL
@@ -1260,13 +1339,13 @@ drop_chMDB <- function(x, name){
       stop("Only chTKCat admin can drop an MDB from ClickHouse")
    }
    con <- x$chcon
-   RClickhouse::dbSendQuery(con, sprintf("DROP DATABASE `%s`", name))
+   DBI::dbSendQuery(con, sprintf("DROP DATABASE `%s`", name))
    ul <- list_chTKCat_users(x) %>% 
       dplyr::filter(!.data$admin)
    if("provider" %in%  colnames(ul)){
       pl <- ul$login[which(ul$provider)]
       if(length(pl) > 0){
-         RClickhouse::dbSendQuery(
+         DBI::dbSendQuery(
             con, 
             sprintf(
                paste(
@@ -1285,7 +1364,7 @@ drop_chMDB <- function(x, name){
       cl <- ul$login
    }
    if(length(cl) > 0){
-      RClickhouse::dbSendQuery(
+      DBI::dbSendQuery(
          con,
          sprintf(
             "REVOKE %s ON `%s`.* FROM %s",
@@ -1340,7 +1419,7 @@ update_chMDB_grants <- function(x, mdb){
    
    ## Revoke read access ----
    if(length(others) > 0){
-      RClickhouse::dbSendQuery(
+      DBI::dbSendQuery(
          con,
          sprintf(
             "REVOKE SELECT ON `%s`.* FROM %s",
@@ -1349,7 +1428,7 @@ update_chMDB_grants <- function(x, mdb){
       )
       modelTables <- names(CHMDB_DATA_MODEL)
       for(tn in modelTables){
-         RClickhouse::dbSendQuery(
+         DBI::dbSendQuery(
             con,
             sprintf(
                "GRANT SELECT ON `%s`.`%s` TO %s",
@@ -1360,7 +1439,7 @@ update_chMDB_grants <- function(x, mdb){
    }
    
    ## Grant read access ----
-   RClickhouse::dbSendQuery(
+   DBI::dbSendQuery(
       con,
       sprintf(
          "GRANT SELECT ON `%s`.* TO %s",
@@ -1370,7 +1449,7 @@ update_chMDB_grants <- function(x, mdb){
    
    ## Revoke write access ----
    if(length(c(readUsers, others)) > 0){
-      RClickhouse::dbSendQuery(
+      DBI::dbSendQuery(
          con,
          sprintf(
             "REVOKE CREATE TABLE, DROP TABLE, ALTER, INSERT ON `%s`.* FROM %s",
@@ -1380,7 +1459,7 @@ update_chMDB_grants <- function(x, mdb){
    }
    
    ## Grant write access ----
-   RClickhouse::dbSendQuery(
+   DBI::dbSendQuery(
       con,
       sprintf(
          paste(
@@ -1701,7 +1780,7 @@ empty_chMDB <- function(
   
    ## Remove tables and re-initialize if necessary ----
    for(tn in toDrop){
-      RClickhouse::dbSendQuery(
+      DBI::dbSendQuery(
          x$chcon,
          sprintf("DROP TABLE `%s`.`%s`", name, tn)
       )
@@ -1979,7 +2058,7 @@ set_chMDB_access <- function(x, mdb, public){
       stop("The database does not exist")
    }
    con <- x$chcon
-   RClickhouse::dbSendQuery(
+   DBI::dbSendQuery(
       con, 
       sprintf(
          "ALTER TABLE `%s`.___Public___ DELETE WHERE 1",
@@ -2123,7 +2202,7 @@ remove_chMDB_user <- function(x, mdb, login){
       stop("The database does not exist")
    }
    con <- x$chcon
-   RClickhouse::dbSendQuery(
+   DBI::dbSendQuery(
       con,
       sprintf(
          "ALTER TABLE `%s`.___MDBUsers___ DELETE WHERE login='%s'",
@@ -2259,7 +2338,7 @@ add_chTKCat_collection <- function(x, json, overwrite=FALSE){
       }
    }
    if(writeCol){
-      RClickhouse::dbSendQuery(
+      DBI::dbSendQuery(
          conn=x$chcon,
          statement=sprintf(
             "ALTER TABLE default.Collections DELETE WHERE title='%s'",
@@ -2303,7 +2382,7 @@ remove_chTKCat_collection <- function(x, title){
    if(!title %in% list_chTKCat_collections(x)$title){
       stop("The collection does not exist in the chTKCat")
    }
-   RClickhouse::dbSendQuery(
+   DBI::dbSendQuery(
       conn=x$chcon,
       statement=sprintf(
          "ALTER TABLE default.Collections DELETE WHERE title='%s'",
@@ -2327,7 +2406,7 @@ collection_members.chTKCat <- function(
 ){
    
    con <- x$chcon
-   dbNames <- DBI::dbGetQuery(con, "SELECT * FROM system.databases") %>% 
+   dbNames <- DBI::dbGetQuery(con, "SELECT name FROM system.databases") %>% 
       dplyr::pull("name") %>% 
       setdiff(CH_RESERVED_DB)