+++ R/namespace.R
@@ -2,13 +2,13 @@
 #' @import dplyr
 #' @import ReDaMoR
 #' @import DBI
-#' @import RClickhouse
 #' @import shiny
 #' @import shinydashboard
 #' @import visNetwork
 #' @import future
 #' @import promises
 #' 
+#' @importFrom ClickHouseHTTP ClickHouseHTTP
 #' @importFrom rlang .data enquo
 #' @importFrom tidyselect eval_select vars_pull eval_rename
 #'